requirements.txt
beautifulsoup4==4.11.2
load-dotenv==0.1.0
lxml==4.9.2
psycopg2==2.9.5
python-telegram-bot==13.7
requests==2.28.2